Update deploy-wazuh-windows-agent-suite.ps1
Corrected syntax and command errors in the new logic I put in the script.  Was failing to pull current group memberships.  This was successfully tested on two systems with currently connected agents with multiple group memberships and a system with a single group membership with and without group membership changes.