Updating check for Wazuh version
In 4.something, Wazuh removed /etc/ossec-init.conf and so the method for checking for the installed version of Wazuh agent changed.  This change addresses this change while leaving the old method in place as an alternate.  I have tested this on 4.x installations and on pre-4.x and the new check works as expected.