Update siem-agent-deploy.ps1
Added the new ways of getting version, by .txt, by commandline or by code and tested precedence. 
-Install now preserves registration unless the agent was not connected to a manager, yet does a complete uninstall and reinstall of Wazuh. 
Removed the unneeded checks, per our discussion. 
If connected but groups change, then it just re-registers without uninstalling/reinstalling the agent.
I also changed the logic where it runs through the checkscript every time, even with -Install...in this later case, only to get the current connection status and the current groups for just in case we can preserve the registration. 
I also removed the same manager check, so if we are switching managers when testing, we have to -uninstall and then install, otherwise it tries to preserve the registration (client.keys) from the wrong manager.