major work
Basic testing shows uninstalls and installs working for MacOS.  We still need to retest Linux use of the script to confirm that none of the MacOS compatibility measures have introduced bugs on the Linux.  We also need to build the MacOS wpk Osquery mgt and installer packages.  Sane ULS log collection also needs to be established since specific log types and filters are required to focus in on worthwhile log collection vs absolute log inundation.